RING-29531 Move stream_chunk.js content into split.js
