+++ lib/http_put.js
@@ -5,7 +5,7 @@ const ecstream = require('ecstream');
 const protocol = require('./protocol');
 const httpUtils = require('./http_utils');
 const utils = require('./utils');
-const { chunkedStreamDemux } = require('./stream_chunk');
+const { chunkedStreamDemux } = require('./split');
 
 /**
  * Demultiplex a stream for replication
@@ -86,7 +86,7 @@ function decidePUT(opContext) {
      * safe side.
      */
     const perChunkMinValid = opContext.fragments.code === 'CP' ?
-              Math.ceil((opContext.fragments.nDataParts + 1)/ 2) :
+              Math.ceil((opContext.fragments.nDataParts + 1) / 2) :
               opContext.fragments.nDataParts + 1;
 
     const success = opContext.status.every(
